Added conditional definition of MODULE_SCOPE for cases where it does not exist (Building basekits, tclkits, with Thread directly compiled in). This required modification of some header inclusions as well, to have the definition in one place instead of about four.
Added conditional definition of Tcl_GetErrorLine for the same
reason. Under static build -DTCL_TIP285 is not usable, as
USE_TCL_STUBS will be missing. And the TGEL changes are TIP 336 also,
i.e. unrelated.

And lastly added conditional definition of the package version, again
for cases where it is missing due to Thread configure not invoked.